Doing "ReactApp - Tic Tac Toe": - Removed the unnecessary page: Logout - Minor styling tweaks
