+++ react/react-training/src/app/screens/Dashboard/index.js
@@ -4,7 +4,6 @@ import { Switch, Route } from 'react-router';
 import { ROUTES } from '@constants/routes.js';
 
 import Game from './screens/Game';
-import Logout from './screens/Logout';
 import Profile from './screens/Profile';
 import TopBar from './screens/Game/components/TopBar';
 
@@ -16,7 +15,6 @@ function Dashboard() {
         <Switch>
           <Route exact path={ROUTES.GAME} component={Game} />
           <Route exact path={ROUTES.PROFILE} component={Profile} />
-          <Route exact path={ROUTES.LOGOUT} component={Logout} />
         </Switch>
       </Fragment>
     </Router>
