--- react/react-training/src/app/screens/Dashboard/screens/Logout/styles.scss
@@ -1,8 +1,0 @@
-@import 'variables/colors';
-@import 'variables/fonts';
-
-.title {
-  color: $blue;
-  font-size: $title-size;
-  text-align: center;
-}
