+++ react/react-training/src/scss/variables/_fonts.scss
@@ -5,7 +5,5 @@ $font-main: 'Roboto', Helvetica, sans-serif;
 $title-size: 25px;
 $boxes-font-size: 15px;
 $button-font-size: 15px;
-
 $top-bar-size: 12px;
-
 $sub-title-size: 18px;
